% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.radiation.functions.R
\name{brightness.temperature.total}
\alias{brightness.temperature.total}
\title{Brightness temperature}
\usage{
brightness.temperature.total(I, epsilon = 1)
}
\arguments{
\item{I}{total radiation energy density in [W*m-2]}

\item{epsilon}{emissivity, in the range of c(0,1)}
}
\value{
The brightness temperature in Kelvin.
}
\description{
Calculate the brightness temperature from the total emitted radiation energy density.
}
\seealso{
\code{\link{total.radiation}} for the opposite.
}

