% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.units.functions.R
\name{cel2kel}
\alias{cel2kel}
\title{Convert temperature in Celsius to Kelvin scale}
\usage{
cel2kel(T)
}
\arguments{
\item{T}{temperature in degrees Celsius}
}
\value{
temperature converted to Kelvin scale
}
\description{
Convert temperature in Celsius to Kelvin scale
}
\details{
Unphysical input temperatures will be replaced by \code{NA}.
}
\seealso{
\code{\link{kel2cel}} for the opposite.
}

