% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.constants.R
\docType{data}
\name{celsius0}
\alias{celsius0}
\title{Zero degrees Celsius scale in Kelvin}
\format{\code{273.15 [K]}}
\usage{
celsius0
}
\description{
Zero degrees Celsius scale in Kelvin
}
\details{
0 degrees C = 273.15 K
}
\keyword{datasets}

