% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.units.functions.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radiant}
\usage{
deg2rad(x)
}
\arguments{
\item{x}{a numeric vector of angles in degrees}
}
\value{
a numeric vector of respective angles in radiant.
}
\description{
Convert angles given in degrees to radiant.
}
\seealso{
\code{\link{rad2deg}} for the opposite.
}

