% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.functions.R
\name{e.magnus}
\alias{e.magnus}
\title{Magnus formula for saturation water vapor pressure}
\source{
\url{https://de.wikipedia.org/wiki/Sättigungsdampfdruck}
}
\usage{
e.magnus(T, over = "water")
}
\arguments{
\item{T}{temperature, either in Kelvin.}

\item{over}{Calculate the saturation vapor pressure over "water" or over
"ice"?}
}
\value{
The saturation water vapour pressure e_s(T) in Pascal [Pa].
}
\description{
Calculate the saturation water vapor pressure e_s(T) either over
 water or ice using the Magnus approximation.
}

