% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.units.functions.R
\name{kel2cel}
\alias{kel2cel}
\title{Convert temperature in Kelvin to Celsius scale}
\usage{
kel2cel(T)
}
\arguments{
\item{T}{temperature in degrees Kelvin}
}
\value{
temperature converted to Celsius scale
}
\description{
Convert temperature in Kelvin to Celsius scale
}
\details{
Negative input temperatures will be replaced by \code{NA}.
}
\seealso{
\code{\link{cel2kel}} for the opposite.
}

