% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.radiation.functions.R
\name{planck.function}
\alias{planck.function}
\title{Create a Planck's law function}
\usage{
planck.function(frequency = NULL, wavelength = NULL, epsilon = NULL,
  T = NULL)
}
\arguments{
\item{frequency}{logical or numeric. Return a Planck's law function for frequency?}

\item{wavelength}{logical or numeric. Return a Planck's law function for wavelength? 
Unused if both \code{frequency} and \code{wavelength} are given.}

\item{epsilon}{numeric. If possible, set the constant emissivity \code{epsilon} for the function. In the range c(0,1).}

\item{T}{numeric. If possible, set the constant Temperature \code{T} for the function. In the range c(0,1).}
}
\value{
a Planck's law function
}
\description{
Create a Planck's law function and possibly set parameters.
}
\details{
The function is compiled to bytecode with \code{cmpfun} from the \code{compiler} package
 if possible.\cr\cr
 If the package \code{functionutils} is available, all given numeric arguments are
 substituted into the returned function.
}
\examples{
planck.function(frequency = T)
# function(frequency,T,epsilon) 
#    epsilon * ( 2 * meteorology::h * frequency ^ 3 ) / meteorology::c ^ 2 * 1 / ( exp( (meteorology::h * frequency ) / ( meteorology::k * T ) ) - 1 )

planck.function(frequency = T, epsilon = 0.5, T = 300)
# function (frequency) 
#	  0.5 * (2 * meteorology::h * frequency^3)/meteorology::c^2 * 1/(exp((meteorology::h * frequency)/(meteorology::k * 300)) - 1)

planck.function(frequency = 3e9)
# function (T, epsilon) 
#    epsilon * (2 * meteorology::h * 3e+09^3)/meteorology::c^2 * 1/(exp((meteorology::h * 3e+09)/(meteorology::k * T)) - 1)
}

