% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.radiation.functions.R
\name{planck}
\alias{planck}
\title{Planck's law}
\source{
\url{https://en.wikipedia.org/wiki/Planck\%27s_law}
}
\usage{
planck(frequency = NULL, wavelength = NULL, T, epsilon = 1)
}
\arguments{
\item{frequency}{frequency in [Hz]}

\item{wavelength}{wavelength in [m] (ignored if \code{frequency} is set)}

\item{epsilon}{emissivity, in the range c(0,1)}

\item{temp}{temperature in [K]}
}
\value{
Spectral radiance of body with emissivity e at absolute temperature T
 and frequency \code{frequency} or rather wavelength \code{wavelength}. The unit 
 is [W*m-2*sr-1*Hz] if \code{frequency} was given or [W*m-2*sr-1*m-1] if 
 wavelength \code{wavelength} was given.
}
\description{
Calculate the spectral radiance of gray body per unit area of 
	emitting surface, per unit solid angle and per spectral unit (frequency or 
 wavelength) using Planck's law.
}
\details{
If both \code{frequency} and wavelength \code{wavelength} are given, a 
 warning will be printed and the frequency taken.
}
\seealso{
\code{\link{brightness.temperature}} for backwards determination of 
 the brightness temperature by planck's law.
}

