% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.radiation.functions.R
\name{position.max.radiation}
\alias{position.max.radiation}
\title{Locate position of maximal radiation}
\source{
\url{https://de.wikipedia.org/wiki/Wiensches_Verschiebungsgesetz}
}
\usage{
position.max.radiation(T, frequency = F, wavelength = T)
}
\arguments{
\item{T}{temperature in Kelvin.}

\item{frequency}{logical. Return the position as frequency? Defaults to FALSE.}

\item{wavelength}{logical. Return the position as wavelength? Defaults to TRUE.}
}
\value{
The position of the maximal radiation at temperature T, either in Hertz [Hz] 
 for \code{frequency} or in metres [m] for \code{wavelength}.
}
\description{
Using Wien's law, determine the frequency or wavelength
 with the maximal radiation at temperature T.
}
\details{
By default, If both \code{frequency} and \code{wavelength} are \code{TRUE},
 the \code{wavelength} is returned.
}

