% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.units.functions.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radiant to degrees}
\usage{
rad2deg(x)
}
\arguments{
\item{x}{a numeric vector of angles in radiant}
}
\value{
a numeric vector of respective angles in degrees.
}
\description{
Convert angles given in radiant to degrees.
}
\seealso{
\code{\link{deg2rad}} for the opposite.
}

