% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.constants.R
\docType{data}
\name{Rdryair}
\alias{Rdryair}
\title{Specific gas constant of dry air}
\format{\code{287.102 [J*kg-1*K-1]}}
\usage{
Rdryair
}
\description{
Specific gas constant of dry air
}
\keyword{datasets}

