% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.statistical.functions.R
\name{rmse}
\alias{rmse}
\title{Root mean square error}
\usage{
rmse(x, y, na.rm = T)
}
\arguments{
\item{x, y}{numerical vectors of same length.}

\item{na.rm}{Remove NAs prior to calculation? Defaults to TRUE.}
}
\description{
Calculate the root mean square error of vectors \code{x}
 and \code{y}.
}
\details{
\code{sqrt( mean( ( x - y ) ^ 2 ) )}
}
\examples{
rmse(1:10,21:30)
# [1] 20
set.seed(42) # We want to produce the same results!
rmse(rnorm(10),rnorm(10))
# [1] 2.108466
}

