% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.constants.R
\docType{data}
\name{Rvapour}
\alias{Rvapour}
\title{Specific gas constant of water vapour}
\format{\code{461.401 [J*kg-1*K-1]}}
\usage{
Rvapour
}
\description{
Specific gas constant of water vapour
}
\keyword{datasets}

