% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteorology.radiation.functions.R
\name{total.radiation}
\alias{total.radiation}
\title{Total body radiation at temperature T after Stefan-Boltzmann.}
\usage{
total.radiation(T, epsilon = 1)
}
\arguments{
\item{T}{temperature in Kelvin [K]}

\item{epsilon}{emissivity, in the range of c(0,1). Defaults to black body \code{epsilon=1}.}
}
\value{
The total body radiation energy density of a body with emissivity 
 \code{epsilon} and temperature \code{T} after Stefan-Boltzmann in [W*m-2].
}
\description{
The total body radiation of a body with emissivity \code{epsilon} and temperature 
 \code{T} after Stefan-Boltzmann.
}
\seealso{
\code{\link{brightness.temperature.total}} for the opposite.
}

